Backing up, backing up... Remove button disabled state.
After discussing with a UX designer, disabled state for button is something that shouldn't really happen. Many users don't understand that the button can't be clicked and think the app is broken. Instead, the buttons should always be clickable and an error message should be shown if the requiremenets for clicking the button isn't met